VR 362 hangup connections (#342)
* event handler catches did rotation event

* add handling for connectiondidrotated

* boolean delete parameter in endpoint

* Bump to credo-ts 0.5.17

* template literals for webhookurl

* pass all DID rotation events

* use @Query

* webhook will always pass connectionRecord

* update tests

+++ src/controllers/v1/connections/ConnectionController.ts
@@ -114,18 +114,41 @@ export class ConnectionController extends Controller {
   }
 
   /**
-   * Deletes a connection record from the connection repository.
-   *
+   * Hangs up an active connection
+   * Optional boolean value to also delete the connection record (default = false)
+   * i.e. /connectionId?deleteConnectionRecord=true
    * @param connectionId Connection identifier
    */
   @Delete('/:connectionId')
+  @Response<BadRequest['message']>(400)
   @Response<NotFoundError['message']>(404)
   @Response<HttpResponse>(500)
-  public async deleteConnection(@Request() req: express.Request, @Path('connectionId') connectionId: UUID) {
+  public async closeConnection(
+    @Request() req: express.Request,
+    @Path('connectionId') connectionId: UUID,
+    @Query('deleteConnectionRecord') deleteConnectionRecord?: boolean
+  ) {
     try {
+      const connectionRecord = await this.agent.connections.getById(connectionId)
+      // If we've hung up on them already, Did will be blank
+      // If they've hung up on us already, theirDid will be blank
+      const alreadyDisconnected = !connectionRecord.theirDid || !connectionRecord.did
+      const deleteAfter = Boolean(deleteConnectionRecord)
+      if (alreadyDisconnected) {
+        if (!deleteAfter) {
+          throw new BadRequest(`cannot send hangup to disconnected peer ${connectionId}`)
+        }
+        await this.agent.connections.deleteById(connectionId)
+        req.log.info('connection record deleted %s', connectionId)
+        this.setStatus(204)
+        return
+      }
+      await this.agent.connections.hangup({ connectionId: connectionId, deleteAfterHangup: deleteAfter })
+      req.log.info(
+        { connectionId, deleteAfterHangup: deleteAfter },
+        deleteAfter ? 'disconnected and deleted' : 'disconnected'
+      )
       this.setStatus(204)
-      await this.agent.connections.deleteById(connectionId)
-      req.log.info('%s connection has been deleted', connectionId)
     } catch (error) {
       if (error instanceof RecordNotFoundError) {
         throw new NotFoundError('connection record not found')
